refactored mock APIs
