+++ test/api/lib/data.json
@@ -1,0 +1,10 @@
+[
+  {
+    "title": "The Prestige",
+    "director": "Christopher Nolan"
+  },
+  {
+    "title": "Schindler's List",
+    "director": "Steven Spielberg"
+  }
+]
